inline dict
